Fix RenderTexture causing crash on exiting program (don't destroy if there's no context). Modified premake files to hide external library warnings (speeds up build times). Fix DuinEditor nfd /MTd.

+++ Duin/src/Duin/Physics/physx/CharacterBody.cpp
@@ -83,7 +83,7 @@ namespace duin {
         Vector3 newPos = Vector3(pxController->getPosition());
         Vector3 distanceMoved = Vector3Subtract(newPos, oldPos);
         
-        currentVelocity = Vector3Scale(distanceMoved, 1.0f / timeSinceLastMove);
+        currentVelocity = Vector3Scale(distanceMoved, (float)(1.0 / timeSinceLastMove));
 
         OnFloorShapeCast(timeSinceLastMove);
